[HB] Exit early if page contains no data
We used to have this logic already for this scraper but
it got lost with the last merge. Note that in this commit
we return None instead of raising a `DropItem` (as we did
in the initial implementation) because we want to silently
drop this item as opposed to treating it like an error.